Migrated tests and benches to the SlotId / u32-capacity API: replaced bare usize slot identifiers with SlotId across all writer/reader/handle accessor surfaces, switched Option<SlotId> comparisons to is_none() / Some(slot) idioms, and updated capacity literals from usize to u32 for the constructors that flipped (NetworkConfig.{node,synapse}_capacity, EntryStoreConfig.capacity, RingBuffer / SimpleFreeList / SlotAllocator / StagingBufferWriter / TripleBufferWriter ::new).
Slot transport via mem_metadata for cross-thread tests now uses slot.to_i32() on the producer side and SlotId::from_i32(...) on the consumer side. Property-test strategies that generated raw usize slots now generate SlotId via SlotId::new(n).unwrap(). Drop-ordering invariant preserved across the migration: where Kernel + EpochConsumer are co-resident, consumer is declared after kernel so it drops first.

+++ packages/synaptic-kernel/tests/common/mod.rs
@@ -40,8 +40,8 @@ pub fn kernel_config_1_1_network(
 /// `dummy_store_capacity` keeps the registry layout valid; tests that never touch the user store
 /// can use a small positive value.
 pub fn kernel_config_1_1(
-    node_capacity: usize,
-    synapse_capacity: usize,
+    node_capacity: u32,
+    synapse_capacity: u32,
     node_meta_stride: usize,
     node_attr_stride: usize,
     synapse_meta_stride: usize,
@@ -60,8 +60,8 @@ pub fn kernel_config_1_1(
 }
 
 pub fn kernel_config_1_1_with_tb(
-    node_capacity: usize,
-    synapse_capacity: usize,
+    node_capacity: u32,
+    synapse_capacity: u32,
     node_meta_stride: usize,
     node_attr_stride: usize,
     synapse_meta_stride: usize,
@@ -82,8 +82,8 @@ pub fn kernel_config_1_1_with_tb(
 
 pub fn kernel_config_1_1_full(
     mem_metadata_size: usize,
-    node_capacity: usize,
-    synapse_capacity: usize,
+    node_capacity: u32,
+    synapse_capacity: u32,
     node_meta_stride: usize,
     node_attr_stride: usize,
     synapse_meta_stride: usize,
